requirements.txt
kafka-python==2.0.2
confluent-kafka==2.3.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
boto3>=1.28.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.21.0

# Note: fire-detect-nn is installed automatically via scripts/install_fire_detect_nn.py
# Run: python3 scripts/install_fire_detect_nn.py
# Or install the package: pip install -e .  (which runs the install script)
